requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleo==2.0.1
coverage==7.1.0
crashtest==0.4.1
cryptography==39.0.0
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.5
django-appconf==1.0.5
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==22.3
django-extensions==3.2.1
django-filter==22.1
django-redis==5.2.0
django-select2==8.0.0
dulwich==0.20.50
exceptiongroup==1.1.0
filelock==3.9.0
flake8==6.0.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
mccabe==0.7.0
more-itertools==9.0.0
msgpack==1.0.4
-e git+ssh://git@github.com/DzmitrySha/ndt_manager.git@2774e602d6360cc10acb3212c2f48486a61bd4e1#egg=ndt_manager
packaging==23.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.2.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pyflakes==3.0.1
pyrsistent==0.19.3
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-dotenv==0.21.1
pytz==2022.1
rapidfuzz==2.13.7
redis==4.4.2
requests==2.28.2
requests-toolbelt==0.10.1
rollbar==0.16.3
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
trove-classifiers==2023.1.20
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.17.1
webencodings==0.5.1
whitenoise==6.3.0
zipp==3.11.0
